+++ src/store/movieMapper.ts
@@ -3,6 +3,6 @@ import { Movie } from "../../types/movie.ts";
 export const mapToMovie = (apiData: any): Movie => ({
   id: apiData.id,
   title: apiData.title,
-  voteAverage: Number(apiData.vote_average.toFixed(1)),
-  posterPath: apiData.poster_path,
+  rating: Number(apiData.vote_average.toFixed(1)),
+  imageSrc: apiData.poster_path,
 });
